requirements.txt
# SEEK Property Platform - Python Dependencies
# Backend data processing and Supabase integration
# Status: Phase 2 - FOIA Integration (Tasks 1-2 Complete ✅, Task 3.2 API Complete ✅)

# Core Dependencies
supabase==2.17.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1

# Data Processing
pandas==2.3.1
numpy==2.3.2
duckdb==1.3.2

# Text Processing and Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
usaddress==0.5.15
us==3.2.0

# File Handling
openpyxl==3.1.5
xlrd==2.0.2

# Development and Testing
pytest==8.3.2
pytest-cov==5.0.0
black==24.8.0
ruff==0.8.5
bandit==1.7.5
safety==3.2.5

# Structured Logging
structlog==25.4.0

# Performance Monitoring
psutil==6.0.0
memory-profiler==0.61.0